Moved screens and viewmodels to subdirectories
