+++ XMRpos/app/src/main/java/org/monerokon/xmrpos/ui/NavGraphRoot.kt
@@ -23,12 +23,12 @@ import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
 import androidx.navigation.toRoute
 import kotlinx.serialization.Serializable
-import org.monerokon.xmrpos.ui.payment.PaymentCheckoutScreenRoot
-import org.monerokon.xmrpos.ui.payment.PaymentCheckoutViewModel
-import org.monerokon.xmrpos.ui.payment.PaymentEntryScreenRoot
-import org.monerokon.xmrpos.ui.payment.PaymentEntryViewModel
-import org.monerokon.xmrpos.ui.payment.PaymentSuccessScreenRoot
-import org.monerokon.xmrpos.ui.payment.PaymentSuccessViewModel
+import org.monerokon.xmrpos.ui.payment.checkout.PaymentCheckoutScreenRoot
+import org.monerokon.xmrpos.ui.payment.checkout.PaymentCheckoutViewModel
+import org.monerokon.xmrpos.ui.payment.entry.PaymentEntryScreenRoot
+import org.monerokon.xmrpos.ui.payment.entry.PaymentEntryViewModel
+import org.monerokon.xmrpos.ui.payment.success.PaymentSuccessScreenRoot
+import org.monerokon.xmrpos.ui.payment.success.PaymentSuccessViewModel
 import org.monerokon.xmrpos.ui.settings.companyinformation.CompanyInformationScreenRoot
 import org.monerokon.xmrpos.ui.settings.companyinformation.CompanyInformationViewModel
 import org.monerokon.xmrpos.ui.settings.exporttransactions.ExportTransactionsScreenRoot
